+++ project/src/components/app/app.tsx
@@ -1,13 +1,15 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import { AppRoute } from '../../constants';
+
 import MainScreen from '../../pages/main-screen/main-screen';
 import LoginScreen from '../../pages/login-screen/login-screen';
 import OfferScreen from '../../pages/offer-screen/offer-screen';
 import FavoritesScreen from '../../pages/favorites-screen/favorites-screen';
 import Page404Screen from '../../pages/page-404-screen/page-404-screen';
 import PrivateRoute from '../private-route/private-route';
-import {Offer} from '../../types/offer';
-import {Review} from '../../types/review';
+import { Offer } from '../../types/offer';
+import { Review } from '../../types/review';
+
+import { AppRoute } from '../../constants';
 
 type AppProps = {
   placeCount: number;
@@ -32,7 +34,10 @@ function App({ placeCount, offers, reviews }: AppProps): JSX.Element {
             </PrivateRoute>
           }
         />
-        <Route path={AppRoute.Offer} element={<OfferScreen reviews={reviews} />} />
+        <Route
+          path={AppRoute.Offer}
+          element={<OfferScreen reviews={reviews} />}
+        />
         <Route path="*" element={<Page404Screen />} />
       </Routes>
     </BrowserRouter>
