+++ project/src/components/private-route/private-route.tsx
@@ -1,4 +1,5 @@
 import { Navigate } from 'react-router-dom';
+
 import { AppRoute } from '../../constants';
 
 type PrivateRouteProps = {
